+++ jabgui/src/main/java/org/jabref/gui/preferences/entryeditor/EntryEditorTab.java
@@ -30,6 +30,7 @@ public class EntryEditorTab extends AbstractPreferenceTabView<EntryEditorTabView
     @FXML private CheckBox journalPopupEnabled;
     @FXML private CheckBox autoLinkFilesEnabled;
     @FXML private CheckBox enableSciteTab;
+    @FXML private CheckBox enableCitationContextTab;
     @FXML private CheckBox showUserCommentsField;
 
     @FXML private Button generalFieldsHelp;
@@ -62,6 +63,7 @@ public class EntryEditorTab extends AbstractPreferenceTabView<EntryEditorTabView
         journalPopupEnabled.selectedProperty().bindBidirectional(viewModel.journalPopupProperty());
         autoLinkFilesEnabled.selectedProperty().bindBidirectional(viewModel.autoLinkFilesEnabledProperty());
         enableSciteTab.selectedProperty().bindBidirectional(viewModel.enableSciteTabProperty());
+        enableCitationContextTab.selectedProperty().bindBidirectional(viewModel.enableCitationContextTabProperty());
         showUserCommentsField.selectedProperty().bindBidirectional(viewModel.showUserCommentsProperty());
 
         fieldsTextArea.textProperty().bindBidirectional(viewModel.fieldsProperty());